feat: Avoid wiping the toolhead on the model when resuming
When resuming the print the toolhead had a chance of wiping plastic on
the outside of the model as it returned.  This patch adds a move back to
a point above where it started before resuming.  With a suitable "dz"
this elminiated branching (like stringing) for me.

Signed off by: Paul Bone <paul@bone.id.au>